+++ Frends.SFTP.DeleteFiles/Frends.SFTP.DeleteFiles/Frends.SFTP.DeleteFiles.cs
@@ -25,8 +25,7 @@ public static class SFTP
     /// <param name="connection">Connection parameters.</param>
     /// <param name="cancellationToken">Cancellation token given by Frends.</param>
     /// <returns>Object { List&lt;FileItem&gt; Files[ Object { string Name, string Path, Double SizeInMegaBytes } ] }.</returns>
-    public static async Task<Result> DeleteFiles([PropertyTab] Input input, [PropertyTab] Connection connection,
-        CancellationToken cancellationToken)
+    public static async Task<Result> DeleteFiles([PropertyTab] Input input, [PropertyTab] Connection connection, CancellationToken cancellationToken)
     {
         var deletedFiles = new List<FileItem>();
 
@@ -76,8 +75,7 @@ public static class SFTP
         return new Result(deletedFiles);
     }
 
-    private static async Task<List<FileItem>> GetFiles(SftpClient sftp, Input input,
-        CancellationToken cancellationToken)
+    private static async Task<List<FileItem>> GetFiles(SftpClient sftp, Input input, CancellationToken cancellationToken)
     {
         var directoryList = new List<FileItem>();
         var filePaths = ConvertObjectToStringArray(input.FilePaths);